refactor(supabase): split end-turn-basic function into focused modules
Keep index.ts as a thin Edge Function entrypoint and split focused helpers into local Edge Function
modules, such as HTTP/CORS, auth, authorization, Supabase REST/RPC clients, payload validation, and
persistence mapping.

fix #119
