+++ supabase/functions/end-turn-basic/index.ts
@@ -1,183 +1,39 @@
-// The @/ path alias is a Vite convention. Supabase Edge Functions cannot
-// resolve it unless an explicit import map is configured for the Edge runtime.
-// Shared pure logic lives under src/shared/ and is imported by relative path.
+import { resolveSupabaseAuthContext } from "./auth.ts";
+import { resolveSupabaseEndTurnAuthorization } from "./authorize.ts";
+import { getEdgeRuntime } from "./env.ts";
+import { corsHeaders, createJsonResponse } from "./http.ts";
+import { persistSupabaseRunningTransition } from "./persist.ts";
+import { resolveSupabaseEndTurnTransitionInput } from "./state.ts";
 import {
-  isBasicEndTurnTransitionPlanningError,
-  planBasicEndTurnTransition,
-} from "../../../src/shared/endTurnTransitionPlanning.ts";
-import {
-  formatCalendarDate,
-  type TurnCalendarConfig,
-} from "../../../src/shared/turnCalendarPrimitives.ts";
+  mapDryWriteTransitionResult,
+  planDryWriteEndTurnTransition,
+} from "./transition.ts";
+import { parseEndTurnBasicRequestBody } from "./validate.ts";
 
-import type {
-  BasicEndTurnReadinessRow,
-  BasicEndTurnTransitionInput,
-  BasicEndTurnTransitionResult,
-} from "../../../src/shared/endTurnTransitionTypes.ts";
+import type { EndTurnBasicHandlerOptions } from "./types.ts";
 
-type EdgeRuntime = {
-  readonly env: {
-    get(name: string): string | undefined;
-  };
-  serve(handler: (request: Request) => Promise<Response> | Response): void;
+export type {
+  EndTurnBasicAuthContext,
+  EndTurnBasicAuthContextResult,
+  EndTurnBasicAuthorizationResult,
+  EndTurnBasicDryWriteTransitionResult,
+  EndTurnBasicErrorCode,
+  EndTurnBasicErrorResponse,
+  EndTurnBasicHandlerOptions,
+  EndTurnBasicPersistedTransition,
+  EndTurnBasicPersistRunningTransitionResult,
+  EndTurnBasicRequestBody,
+  EndTurnBasicResponse,
+  EndTurnBasicSuccessResponse,
+  EndTurnBasicTransitionInputResult,
+} from "./types.ts";
+
+export {
+  persistSupabaseRunningTransition,
+  resolveSupabaseAuthContext,
+  resolveSupabaseEndTurnAuthorization,
+  resolveSupabaseEndTurnTransitionInput,
 };
-
-declare const Deno: EdgeRuntime | undefined;
-
-export type EndTurnBasicRequestBody = {
-  readonly expectedTurnNumber: number;
-  readonly worldId: string;
-};
-
-export type EndTurnBasicErrorCode =
-  | "auth_context_unavailable"
-  | "end_turn_calendar_config_invalid"
-  | "end_turn_running_transition"
-  | "end_turn_stale_expected_turn"
-  | "end_turn_state_unavailable"
-  | "end_turn_transition_failed"
-  | "end_turn_transition_unavailable"
-  | "end_turn_world_archived"
-  | "end_turn_world_not_found"
-  | "invalid_request"
-  | "method_not_allowed"
-  | "unauthorized"
-  | "unauthenticated";
-
-export type EndTurnBasicErrorResponse = {
-  readonly error: {
-    readonly code: EndTurnBasicErrorCode;
-    readonly details?: readonly string[];
-    readonly message: string;
-  };
-  readonly ok: false;
-};
-
-export type EndTurnBasicSuccessResponse = {
-  readonly data: {
-    readonly actorId: string;
-    readonly transition: EndTurnBasicDryWriteTransitionResult;
-    readonly worldId: string;
-  };
-  readonly ok: true;
-};
-
-export type EndTurnBasicResponse =
-  | EndTurnBasicErrorResponse
-  | EndTurnBasicSuccessResponse;
-
-export type EndTurnBasicDryWriteTransitionResult = {
-  readonly nextDate: BasicEndTurnTransitionResult["nextDate"];
-  readonly nextDateLabel: string;
-  readonly nextTurnNumber: number;
-  readonly previousDate: BasicEndTurnTransitionResult["previousDate"];
-  readonly previousDateLabel: string;
-  readonly previousTurnNumber: number;
-  readonly readinessSummary: BasicEndTurnTransitionResult["readinessSummary"];
-};
-
-export type EndTurnBasicAuthContext = {
-  readonly authorizationHeader?: string;
-  readonly userId: string;
-};
-
-export type EndTurnBasicPersistedTransition = {
-  readonly fromTurnNumber: number;
-  readonly id: string;
-  readonly initiatedByUserId: string;
-  readonly startedAt: string;
-  readonly status: "completed" | "failed" | "running";
-  readonly toTurnNumber: number;
-  readonly worldId: string;
-};
-
-export type EndTurnBasicAuthContextResult =
-  | {
-      readonly context: EndTurnBasicAuthContext;
-      readonly ok: true;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-      readonly status: number;
-    };
-
-export type EndTurnBasicAuthorizationResult =
-  | {
-      readonly ok: true;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-      readonly status: number;
-    };
-
-export type EndTurnBasicTransitionInputResult =
-  | {
-      readonly input: BasicEndTurnTransitionInput;
-      readonly ok: true;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-      readonly status: number;
-    };
-
-export type EndTurnBasicPersistRunningTransitionResult =
-  | {
-      readonly ok: true;
-      readonly transition: EndTurnBasicPersistedTransition;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-      readonly status: number;
-    };
-
-export type EndTurnBasicHandlerOptions = {
-  readonly resolveTransitionInput?: (
-    requestBody: EndTurnBasicRequestBody,
-    authContext: EndTurnBasicAuthContext,
-  ) => Promise<EndTurnBasicTransitionInputResult>;
-  readonly resolveAuthorization?: (
-    requestBody: EndTurnBasicRequestBody,
-    authContext: EndTurnBasicAuthContext,
-  ) => Promise<EndTurnBasicAuthorizationResult>;
-  readonly resolveAuthContext?: (
-    request: Request,
-  ) => Promise<EndTurnBasicAuthContextResult>;
-  readonly persistRunningTransition?: (
-    input: BasicEndTurnTransitionInput,
-    transition: BasicEndTurnTransitionResult,
-    authContext: EndTurnBasicAuthContext,
-  ) => Promise<EndTurnBasicPersistRunningTransitionResult>;
-};
-
-const corsHeaders = {
-  "access-control-allow-headers":
-    "authorization, x-client-info, apikey, content-type",
-  "access-control-allow-methods": "POST, OPTIONS",
-  "access-control-allow-origin": "*",
-} as const;
-
-const jsonHeaders = {
-  "content-type": "application/json; charset=utf-8",
-  ...corsHeaders,
-} as const;
-
-const expectedRequestFields = ["expectedTurnNumber", "worldId"] as const;
-const expectedCalendarConfigFields = [
-  "dateFormatTemplate",
-  "months",
-  "startingDayOfMonth",
-  "startingMonthIndex",
-  "startingWeekdayOffset",
-  "startingYear",
-  "weekdays",
-] as const;
-const expectedCalendarMonthFields = ["dayCount", "index", "name"] as const;
-const expectedCalendarWeekdayFields = ["index", "name"] as const;
 
 export async function handleEndTurnBasicRequest(
   request: Request,
@@ -189,10 +45,13 @@ export async function handleEndTurnBasicRequest(
 
   if (request.method !== "POST") {
     return createJsonResponse(
-      createErrorResponse({
-        code: "method_not_allowed",
-        message: "Use POST to request an end-turn transition.",
-      }),
+      {
+        error: {
+          code: "method_not_allowed",
+          message: "Use POST to request an end-turn transition.",
+        },
+        ok: false,
+      },
       405,
     );
   }
@@ -253,9 +112,9 @@ export async function handleEndTurnBasicRequest(
     );
   }
 
-  const persistRunningTransition =
+  const persistTransition =
     options.persistRunningTransition ?? persistSupabaseRunningTransition;
-  const persistedTransitionResult = await persistRunningTransition(
+  const persistedTransitionResult = await persistTransition(
     transitionInputResult.input,
     plannedTransitionResult.transition,
     authContextResult.context,
@@ -282,1363 +141,6 @@ export async function handleEndTurnBasicRequest(
     },
     200,
   );
-}
-
-export async function resolveSupabaseEndTurnTransitionInput(
-  requestBody: EndTurnBasicRequestBody,
-  authContext: EndTurnBasicAuthContext,
-): Promise<EndTurnBasicTransitionInputResult> {
-  const authorizationHeader = authContext.authorizationHeader;
-
-  if (authorizationHeader === undefined) {
-    return createTransitionStateUnavailableResult();
-  }
-
-  const supabaseUrl = getRequiredRuntimeEnv("SUPABASE_URL");
-  const supabaseAnonKey = getRequiredRuntimeEnv("SUPABASE_ANON_KEY");
-
-  if (supabaseUrl === undefined || supabaseAnonKey === undefined) {
-    return createTransitionStateUnavailableResult();
-  }
-
-  const worldResult = await fetchSupabaseEndTurnWorldState({
-    authorizationHeader,
-    supabaseAnonKey,
-    supabaseUrl,
-    worldId: requestBody.worldId,
-  });
-
-  if (!worldResult.ok) {
-    return transitionInputResultFromStateFetchError(worldResult.error);
-  }
-
-  const calendarConfig = parseWorldCalendarConfig(
-    worldResult.row.calendar_config_json,
-  );
-
-  if (calendarConfig === null) {
-    return {
-      error: createErrorResponse({
-        code: "end_turn_calendar_config_invalid",
-        message: "Calendar configuration is invalid.",
-      }),
-      ok: false,
-      status: 500,
-    };
-  }
-
-  const readinessRowsResult = await fetchSupabaseEndTurnReadinessRows({
-    authorizationHeader,
-    supabaseAnonKey,
-    supabaseUrl,
-    worldId: requestBody.worldId,
-  });
-
-  if (!readinessRowsResult.ok) {
-    return transitionInputResultFromStateFetchError(readinessRowsResult.error);
-  }
-
-  return {
-    input: {
-      actorId: authContext.userId,
-      calendarConfig,
-      currentTurnNumber: worldResult.row.current_turn_number,
-      expectedCurrentTurnNumber: requestBody.expectedTurnNumber,
-      isWorldArchived: worldResult.row.status === "archived",
-      readinessRows: readinessRowsResult.rows,
-      worldId: worldResult.row.id,
-    },
-    ok: true,
-  };
-}
-
-export async function persistSupabaseRunningTransition(
-  input: BasicEndTurnTransitionInput,
-  transition: BasicEndTurnTransitionResult,
-  authContext: EndTurnBasicAuthContext,
-): Promise<EndTurnBasicPersistRunningTransitionResult> {
-  const authorizationHeader = authContext.authorizationHeader;
-
-  if (authorizationHeader === undefined) {
-    return createTransitionPersistenceUnavailableResult();
-  }
-
-  const supabaseUrl = getRequiredRuntimeEnv("SUPABASE_URL");
-  const supabaseAnonKey = getRequiredRuntimeEnv("SUPABASE_ANON_KEY");
-
-  if (supabaseUrl === undefined || supabaseAnonKey === undefined) {
-    return createTransitionPersistenceUnavailableResult();
-  }
-
-  if (
-    transition.fromTurnNumber !== input.expectedCurrentTurnNumber ||
-    transition.toTurnNumber !== transition.fromTurnNumber + 1
-  ) {
-    return createTransitionPersistenceUnavailableResult();
-  }
-
-  const advanceResult = await advanceSupabaseWorldTurn({
-    authorizationHeader,
-    expectedTurnNumber: input.expectedCurrentTurnNumber,
-    logPayload: transition.logPayload,
-    notificationPayload: transition.notificationPayload,
-    supabaseAnonKey,
-    supabaseUrl,
-    worldId: input.worldId,
-  });
-
-  if (advanceResult.ok) {
-    return {
-      ok: true,
-      transition: advanceResult.transition,
-    };
-  }
-
-  return transitionPersistenceResultFromFetchError(advanceResult.error);
-}
-
-export async function resolveSupabaseAuthContext(
-  request: Request,
-): Promise<EndTurnBasicAuthContextResult> {
-  const authorizationHeader = getAuthorizationHeader(request);
-
-  if (authorizationHeader === null) {
-    return createAuthErrorResult();
-  }
-
-  const supabaseUrl = getRequiredRuntimeEnv("SUPABASE_URL");
-  const supabaseAnonKey = getRequiredRuntimeEnv("SUPABASE_ANON_KEY");
-
-  if (supabaseUrl === undefined || supabaseAnonKey === undefined) {
-    return {
-      error: createErrorResponse({
-        code: "auth_context_unavailable",
-        message: "Supabase auth configuration is unavailable.",
-      }),
-      ok: false,
-      status: 500,
-    };
-  }
-
-  const authResponse = await fetch(`${supabaseUrl}/auth/v1/user`, {
-    headers: {
-      apikey: supabaseAnonKey,
-      authorization: authorizationHeader,
-    },
-    method: "GET",
-  });
-
-  if (!authResponse.ok) {
-    return createAuthErrorResult();
-  }
-
-  const authPayload: unknown = await authResponse.json();
-
-  if (!isAuthUserPayload(authPayload)) {
-    return createAuthErrorResult();
-  }
-
-  return {
-    context: {
-      authorizationHeader,
-      userId: authPayload.id,
-    },
-    ok: true,
-  };
-}
-
-export async function resolveSupabaseEndTurnAuthorization(
-  requestBody: EndTurnBasicRequestBody,
-  authContext: EndTurnBasicAuthContext,
-): Promise<EndTurnBasicAuthorizationResult> {
-  const authorizationHeader = authContext.authorizationHeader;
-
-  if (authorizationHeader === undefined) {
-    return createAuthContextUnavailableResult();
-  }
-
-  const supabaseUrl = getRequiredRuntimeEnv("SUPABASE_URL");
-  const supabaseAnonKey = getRequiredRuntimeEnv("SUPABASE_ANON_KEY");
-
-  if (supabaseUrl === undefined || supabaseAnonKey === undefined) {
-    return createAuthContextUnavailableResult();
-  }
-
-  const superAdminResult = await fetchSupabaseRpcBoolean({
-    authorizationHeader,
-    body: {},
-    functionName: "is_super_admin",
-    supabaseAnonKey,
-    supabaseUrl,
-  });
-
-  if (!superAdminResult.ok) {
-    return resultFromSupabaseAuthorizationFetchError(superAdminResult.error);
-  }
-
-  if (superAdminResult.value) {
-    const worldExistsResult = await fetchSupabaseWorldExists({
-      authorizationHeader,
-      supabaseAnonKey,
-      supabaseUrl,
-      worldId: requestBody.worldId,
-    });
-
-    if (!worldExistsResult.ok) {
-      return resultFromSupabaseAuthorizationFetchError(worldExistsResult.error);
-    }
-
-    if (worldExistsResult.value) {
-      return { ok: true };
-    }
-
-    return createAuthorizationErrorResult();
-  }
-
-  const worldAdminResult = await fetchSupabaseRpcBoolean({
-    authorizationHeader,
-    body: {
-      p_world_id: requestBody.worldId,
-    },
-    functionName: "is_world_admin",
-    supabaseAnonKey,
-    supabaseUrl,
-  });
-
-  if (!worldAdminResult.ok) {
-    return resultFromSupabaseAuthorizationFetchError(worldAdminResult.error);
-  }
-
-  if (!worldAdminResult.value) {
-    return createAuthorizationErrorResult();
-  }
-
-  return { ok: true };
-}
-
-async function parseEndTurnBasicRequestBody(request: Request): Promise<
-  | {
-      readonly body: EndTurnBasicRequestBody;
-      readonly ok: true;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-    }
-> {
-  let body: unknown;
-
-  try {
-    body = await request.json();
-  } catch {
-    return {
-      error: createErrorResponse({
-        code: "invalid_request",
-        details: ["body"],
-        message: "Request body must be valid JSON.",
-      }),
-      ok: false,
-    };
-  }
-
-  const bodyShapeResult = parseEndTurnBasicRequestBodyShape(body);
-
-  if (!bodyShapeResult.ok) {
-    return {
-      error: createErrorResponse({
-        code: "invalid_request",
-        details: bodyShapeResult.validationErrors,
-        message: "Request body must include worldId and expectedTurnNumber.",
-      }),
-      ok: false,
-    };
-  }
-
-  return {
-    body: {
-      expectedTurnNumber: bodyShapeResult.body.expectedTurnNumber,
-      worldId: bodyShapeResult.body.worldId.trim(),
-    },
-    ok: true,
-  };
-}
-
-function parseEndTurnBasicRequestBodyShape(body: unknown):
-  | {
-      readonly body: EndTurnBasicRequestBody;
-      readonly ok: true;
-    }
-  | {
-      readonly ok: false;
-      readonly validationErrors: readonly string[];
-    } {
-  const validationErrors = validateEndTurnBasicRequestBody(body);
-
-  if (validationErrors.length > 0 || !isEndTurnBasicRequestBody(body)) {
-    return {
-      ok: false,
-      validationErrors,
-    };
-  }
-
-  return {
-    body,
-    ok: true,
-  };
-}
-
-function validateEndTurnBasicRequestBody(body: unknown): readonly string[] {
-  if (!isRecord(body)) {
-    return ["body"];
-  }
-
-  const validationErrors: string[] = [];
-
-  if (!hasOnlyExpectedFields(body, expectedRequestFields)) {
-    validationErrors.push("body");
-  }
-
-  if (typeof body.worldId !== "string" || body.worldId.trim().length === 0) {
-    validationErrors.push("worldId");
-  }
-
-  if (
-    typeof body.expectedTurnNumber !== "number" ||
-    !Number.isSafeInteger(body.expectedTurnNumber) ||
-    body.expectedTurnNumber < 0
-  ) {
-    validationErrors.push("expectedTurnNumber");
-  }
-
-  return validationErrors;
-}
-
-function isEndTurnBasicRequestBody(
-  body: unknown,
-): body is EndTurnBasicRequestBody {
-  return (
-    isRecord(body) &&
-    hasOnlyExpectedFields(body, expectedRequestFields) &&
-    typeof body.worldId === "string" &&
-    body.worldId.trim().length > 0 &&
-    typeof body.expectedTurnNumber === "number" &&
-    Number.isSafeInteger(body.expectedTurnNumber) &&
-    body.expectedTurnNumber >= 0
-  );
-}
-
-function hasOnlyExpectedFields(
-  body: Record<string, unknown>,
-  expectedFields: readonly string[],
-): boolean {
-  return Object.keys(body).every((fieldName) =>
-    expectedFields.some((expectedField) => expectedField === fieldName),
-  );
-}
-
-function createAuthErrorResult(): EndTurnBasicAuthContextResult {
-  return {
-    error: createErrorResponse({
-      code: "unauthenticated",
-      message: "An authenticated Supabase session is required.",
-    }),
-    ok: false,
-    status: 401,
-  };
-}
-
-function createAuthorizationErrorResult(): EndTurnBasicAuthorizationResult {
-  return {
-    error: createErrorResponse({
-      code: "unauthorized",
-      message: "End turn is unavailable for this world.",
-    }),
-    ok: false,
-    status: 403,
-  };
-}
-
-function createAuthContextUnavailableResult(): EndTurnBasicAuthorizationResult {
-  return {
-    error: createErrorResponse({
-      code: "auth_context_unavailable",
-      message: "Supabase auth configuration is unavailable.",
-    }),
-    ok: false,
-    status: 500,
-  };
-}
-
-function createTransitionStateUnavailableResult(): {
-  readonly error: EndTurnBasicErrorResponse;
-  readonly ok: false;
-  readonly status: number;
-} {
-  return {
-    error: createErrorResponse({
-      code: "end_turn_state_unavailable",
-      message: "End turn state is unavailable.",
-    }),
-    ok: false,
-    status: 500,
-  };
-}
-
-function createTransitionPersistenceUnavailableResult(): EndTurnBasicPersistRunningTransitionResult {
-  return {
-    error: createErrorResponse({
-      code: "end_turn_transition_unavailable",
-      message: "End turn transition could not be started.",
-    }),
-    ok: false,
-    status: 500,
-  };
-}
-
-function createRunningTransitionResult(): EndTurnBasicPersistRunningTransitionResult {
-  return {
-    error: createErrorResponse({
-      code: "end_turn_running_transition",
-      message: "Another end-turn transition is already running.",
-    }),
-    ok: false,
-    status: 409,
-  };
-}
-
-function createTransitionFailedResult(): EndTurnBasicPersistRunningTransitionResult {
-  return {
-    error: createErrorResponse({
-      code: "end_turn_transition_failed",
-      message: "End turn persistence failed after the transition started.",
-    }),
-    ok: false,
-    status: 500,
-  };
-}
-
-function planDryWriteEndTurnTransition(input: BasicEndTurnTransitionInput):
-  | {
-      readonly ok: true;
-      readonly transition: BasicEndTurnTransitionResult;
-    }
-  | {
-      readonly error: EndTurnBasicErrorResponse;
-      readonly ok: false;
-      readonly status: number;
-    } {
-  try {
-    return {
-      ok: true,
-      transition: planBasicEndTurnTransition(input),
-    };
-  } catch (error) {
-    if (isBasicEndTurnTransitionPlanningError(error)) {
-      return {
-        error: createErrorResponse({
-          code: error.code,
-          message: error.message,
-        }),
-        ok: false,
-        status: 409,
-      };
-    }
-
-    return createTransitionStateUnavailableResult();
-  }
-}
-
-function mapDryWriteTransitionResult(
-  transition: BasicEndTurnTransitionResult,
-  dateFormatTemplate: string,
-): EndTurnBasicDryWriteTransitionResult {
-  return {
-    nextDate: transition.nextDate,
-    nextDateLabel: formatCalendarDate(transition.nextDate, {
-      dateFormatTemplate,
-    }),
-    nextTurnNumber: transition.toTurnNumber,
-    previousDate: transition.previousDate,
-    previousDateLabel: formatCalendarDate(transition.previousDate, {
-      dateFormatTemplate,
-    }),
-    previousTurnNumber: transition.fromTurnNumber,
-    readinessSummary: transition.readinessSummary,
-  };
-}
-
-function createErrorResponse({
-  code,
-  details,
-  message,
-}: {
-  readonly code: EndTurnBasicErrorCode;
-  readonly details?: readonly string[];
-  readonly message: string;
-}): EndTurnBasicErrorResponse {
-  if (details === undefined) {
-    return {
-      error: {
-        code,
-        message,
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    error: {
-      code,
-      details,
-      message,
-    },
-    ok: false,
-  };
-}
-
-function createJsonResponse(
-  body: EndTurnBasicResponse,
-  status: number,
-): Response {
-  return new Response(JSON.stringify(body), {
-    headers: jsonHeaders,
-    status,
-  });
-}
-
-function getAuthorizationHeader(request: Request): string | null {
-  const authorizationHeader = request.headers.get("authorization");
-
-  if (authorizationHeader === null) {
-    return null;
-  }
-
-  const trimmedHeader = authorizationHeader.trim();
-
-  if (!trimmedHeader.toLowerCase().startsWith("bearer ")) {
-    return null;
-  }
-
-  const bearerToken = trimmedHeader.slice("bearer ".length).trim();
-
-  if (bearerToken.length === 0) {
-    return null;
-  }
-
-  return `Bearer ${bearerToken}`;
-}
-
-function getRequiredRuntimeEnv(name: string): string | undefined {
-  const edgeRuntime = getEdgeRuntime();
-
-  if (edgeRuntime === undefined) {
-    return undefined;
-  }
-
-  const value = edgeRuntime.env.get(name);
-
-  if (value === undefined || value.trim().length === 0) {
-    return undefined;
-  }
-
-  return value.replace(/\/$/, "");
-}
-
-function getEdgeRuntime(): EdgeRuntime | undefined {
-  if (typeof Deno === "undefined") {
-    return undefined;
-  }
-
-  return Deno;
-}
-
-function isAuthUserPayload(value: unknown): value is { readonly id: string } {
-  return isRecord(value) && typeof value.id === "string" && value.id.length > 0;
-}
-
-type SupabaseAuthorizationFetchError = {
-  readonly safeDeny: boolean;
-};
-
-type SupabaseStateFetchError =
-  | {
-      readonly reason: "fetch_failed" | "invalid_payload";
-    }
-  | {
-      readonly reason: "http_error";
-      readonly safeDeny: boolean;
-    }
-  | {
-      readonly reason: "missing_world";
-    };
-
-type SupabaseBooleanFetchResult =
-  | {
-      readonly ok: true;
-      readonly value: boolean;
-    }
-  | {
-      readonly error: SupabaseAuthorizationFetchError;
-      readonly ok: false;
-    };
-
-type SupabaseEndTurnWorldStateRow = {
-  readonly calendar_config_json: unknown;
-  readonly current_turn_number: number;
-  readonly id: string;
-  readonly status: "active" | "archived";
-};
-
-type SupabaseEndTurnWorldStateResult =
-  | {
-      readonly ok: true;
-      readonly row: SupabaseEndTurnWorldStateRow;
-    }
-  | {
-      readonly error: SupabaseStateFetchError;
-      readonly ok: false;
-    };
-
-type SupabaseEndTurnReadinessRowsResult =
-  | {
-      readonly ok: true;
-      readonly rows: readonly BasicEndTurnReadinessRow[];
-    }
-  | {
-      readonly error: SupabaseStateFetchError;
-      readonly ok: false;
-    };
-
-type SupabasePersistedTransitionRow = {
-  readonly from_turn_number: number;
-  readonly id: string;
-  readonly initiated_by_user_id: string;
-  readonly started_at: string;
-  readonly status: "completed" | "failed" | "running";
-  readonly to_turn_number: number;
-  readonly world_id: string;
-};
-
-type SupabaseRunningTransitionFetchError =
-  | {
-      readonly reason:
-        | "fetch_failed"
-        | "invalid_payload"
-        | "running_transition"
-        | "stale_world_turn"
-        | "transition_failed";
-    }
-  | {
-      readonly reason: "http_error";
-      readonly safeDeny: boolean;
-    }
-  | {
-      readonly reason: "missing_transition";
-    };
-
-type SupabaseRunningTransitionResult =
-  | {
-      readonly ok: true;
-      readonly transition: EndTurnBasicPersistedTransition;
-    }
-  | {
-      readonly error: SupabaseRunningTransitionFetchError;
-      readonly ok: false;
-    };
-
-async function fetchSupabaseRpcBoolean({
-  authorizationHeader,
-  body,
-  functionName,
-  supabaseAnonKey,
-  supabaseUrl,
-}: {
-  readonly authorizationHeader: string;
-  readonly body: Record<string, unknown>;
-  readonly functionName: string;
-  readonly supabaseAnonKey: string;
-  readonly supabaseUrl: string;
-}): Promise<SupabaseBooleanFetchResult> {
-  let response: Response;
-
-  try {
-    response = await fetch(`${supabaseUrl}/rest/v1/rpc/${functionName}`, {
-      body: JSON.stringify(body),
-      headers: {
-        apikey: supabaseAnonKey,
-        authorization: authorizationHeader,
-        "content-type": "application/json",
-      },
-      method: "POST",
-    });
-  } catch {
-    return {
-      error: {
-        safeDeny: false,
-      },
-      ok: false,
-    };
-  }
-
-  if (!response.ok) {
-    return {
-      error: {
-        safeDeny: response.status >= 400 && response.status < 500,
-      },
-      ok: false,
-    };
-  }
-
-  const payload: unknown = await response.json();
-
-  if (typeof payload !== "boolean") {
-    return {
-      error: {
-        safeDeny: false,
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    ok: true,
-    value: payload,
-  };
-}
-
-async function fetchSupabaseEndTurnWorldState({
-  authorizationHeader,
-  supabaseAnonKey,
-  supabaseUrl,
-  worldId,
-}: {
-  readonly authorizationHeader: string;
-  readonly supabaseAnonKey: string;
-  readonly supabaseUrl: string;
-  readonly worldId: string;
-}): Promise<SupabaseEndTurnWorldStateResult> {
-  const searchParameters = new URLSearchParams({
-    id: `eq.${worldId}`,
-    limit: "1",
-    select: "id,current_turn_number,status,calendar_config_json",
-  });
-  let response: Response;
-
-  try {
-    response = await fetch(
-      `${supabaseUrl}/rest/v1/worlds?${searchParameters}`,
-      {
-        headers: {
-          apikey: supabaseAnonKey,
-          authorization: authorizationHeader,
-        },
-        method: "GET",
-      },
-    );
-  } catch {
-    return {
-      error: {
-        reason: "fetch_failed",
-      },
-      ok: false,
-    };
-  }
-
-  if (!response.ok) {
-    return {
-      error: {
-        reason: "http_error",
-        safeDeny: response.status >= 400 && response.status < 500,
-      },
-      ok: false,
-    };
-  }
-
-  const payload: unknown = await response.json();
-
-  if (!Array.isArray(payload)) {
-    return {
-      error: {
-        reason: "invalid_payload",
-      },
-      ok: false,
-    };
-  }
-
-  const rows: readonly unknown[] = payload;
-  const row = rows[0];
-
-  if (row === undefined) {
-    return {
-      error: {
-        reason: "missing_world",
-      },
-      ok: false,
-    };
-  }
-
-  if (!isSupabaseEndTurnWorldStateRow(row)) {
-    return {
-      error: {
-        reason: "invalid_payload",
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    ok: true,
-    row,
-  };
-}
-
-async function advanceSupabaseWorldTurn({
-  authorizationHeader,
-  expectedTurnNumber,
-  logPayload,
-  notificationPayload,
-  supabaseAnonKey,
-  supabaseUrl,
-  worldId,
-}: {
-  readonly authorizationHeader: string;
-  readonly expectedTurnNumber: number;
-  readonly logPayload: BasicEndTurnTransitionResult["logPayload"];
-  readonly notificationPayload: BasicEndTurnTransitionResult["notificationPayload"];
-  readonly supabaseAnonKey: string;
-  readonly supabaseUrl: string;
-  readonly worldId: string;
-}): Promise<SupabaseRunningTransitionResult> {
-  let response: Response;
-
-  try {
-    response = await fetch(
-      `${supabaseUrl}/rest/v1/rpc/advance_world_turn_if_current`,
-      {
-        body: JSON.stringify({
-          p_expected_turn_number: expectedTurnNumber,
-          p_log_payload_jsonb: logPayload,
-          p_notification_payload_jsonb: notificationPayload,
-          p_world_id: worldId,
-        }),
-        headers: {
-          apikey: supabaseAnonKey,
-          authorization: authorizationHeader,
-          "content-type": "application/json",
-        },
-        method: "POST",
-      },
-    );
-  } catch {
-    return {
-      error: {
-        reason: "fetch_failed",
-      },
-      ok: false,
-    };
-  }
-
-  if (!response.ok) {
-    return {
-      error: {
-        reason: "http_error",
-        safeDeny: response.status >= 400 && response.status < 500,
-      },
-      ok: false,
-    };
-  }
-
-  return advanceWorldTurnResultFromResponse(response);
-}
-
-async function fetchSupabaseEndTurnReadinessRows({
-  authorizationHeader,
-  supabaseAnonKey,
-  supabaseUrl,
-  worldId,
-}: {
-  readonly authorizationHeader: string;
-  readonly supabaseAnonKey: string;
-  readonly supabaseUrl: string;
-  readonly worldId: string;
-}): Promise<SupabaseEndTurnReadinessRowsResult> {
-  const searchParameters = new URLSearchParams({
-    "nations.world_id": `eq.${worldId}`,
-    order: "id.asc",
-    select: "id,auto_ready_enabled,is_ready_current_turn,nations!inner()",
-  });
-  let response: Response;
-
-  try {
-    response = await fetch(
-      `${supabaseUrl}/rest/v1/settlements?${searchParameters}`,
-      {
-        headers: {
-          apikey: supabaseAnonKey,
-          authorization: authorizationHeader,
-        },
-        method: "GET",
-      },
-    );
-  } catch {
-    return {
-      error: {
-        reason: "fetch_failed",
-      },
-      ok: false,
-    };
-  }
-
-  if (!response.ok) {
-    return {
-      error: {
-        reason: "http_error",
-        safeDeny: response.status >= 400 && response.status < 500,
-      },
-      ok: false,
-    };
-  }
-
-  const payload: unknown = await response.json();
-
-  if (!Array.isArray(payload) || !payload.every(isSupabaseReadinessRow)) {
-    return {
-      error: {
-        reason: "invalid_payload",
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    ok: true,
-    rows: payload.map(toBasicEndTurnReadinessRow),
-  };
-}
-
-async function fetchSupabaseWorldExists({
-  authorizationHeader,
-  supabaseAnonKey,
-  supabaseUrl,
-  worldId,
-}: {
-  readonly authorizationHeader: string;
-  readonly supabaseAnonKey: string;
-  readonly supabaseUrl: string;
-  readonly worldId: string;
-}): Promise<SupabaseBooleanFetchResult> {
-  const searchParameters = new URLSearchParams({
-    id: `eq.${worldId}`,
-    limit: "1",
-    select: "id",
-  });
-  let response: Response;
-
-  try {
-    response = await fetch(
-      `${supabaseUrl}/rest/v1/worlds?${searchParameters}`,
-      {
-        headers: {
-          apikey: supabaseAnonKey,
-          authorization: authorizationHeader,
-        },
-        method: "GET",
-      },
-    );
-  } catch {
-    return {
-      error: {
-        safeDeny: false,
-      },
-      ok: false,
-    };
-  }
-
-  if (!response.ok) {
-    return {
-      error: {
-        safeDeny: response.status >= 400 && response.status < 500,
-      },
-      ok: false,
-    };
-  }
-
-  const payload: unknown = await response.json();
-
-  if (!Array.isArray(payload)) {
-    return {
-      error: {
-        safeDeny: false,
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    ok: true,
-    value: payload.length > 0,
-  };
-}
-
-async function runningTransitionResultFromResponse(
-  response: Response,
-): Promise<SupabaseRunningTransitionResult> {
-  const payload: unknown = await response.json();
-
-  if (!Array.isArray(payload)) {
-    return {
-      error: {
-        reason: "invalid_payload",
-      },
-      ok: false,
-    };
-  }
-
-  const rows: readonly unknown[] = payload;
-  const row = rows[0];
-
-  if (row === undefined) {
-    return {
-      error: {
-        reason: "missing_transition",
-      },
-      ok: false,
-    };
-  }
-
-  if (!isSupabasePersistedTransitionRow(row)) {
-    return {
-      error: {
-        reason: "invalid_payload",
-      },
-      ok: false,
-    };
-  }
-
-  return {
-    ok: true,
-    transition: toEndTurnBasicPersistedTransition(row),
-  };
-}
-
-async function advanceWorldTurnResultFromResponse(
-  response: Response,
-): Promise<SupabaseRunningTransitionResult> {
-  const result = await runningTransitionResultFromResponse(response);
-
-  if (!result.ok && result.error.reason === "missing_transition") {
-    return {
-      error: {
-        reason: "stale_world_turn",
-      },
-      ok: false,
-    };
-  }
-
-  if (result.ok && result.transition.status === "failed") {
-    return {
-      error: {
-        reason: "transition_failed",
-      },
-      ok: false,
-    };
-  }
-
-  if (result.ok && result.transition.status === "running") {
-    return {
-      error: {
-        reason: "running_transition",
-      },
-      ok: false,
-    };
-  }
-
-  return result;
-}
-
-function transitionPersistenceResultFromFetchError(
-  error: SupabaseRunningTransitionFetchError,
-): EndTurnBasicPersistRunningTransitionResult {
-  if (error.reason === "stale_world_turn") {
-    return {
-      error: createErrorResponse({
-        code: "end_turn_stale_expected_turn",
-        message: "Expected current turn no longer matches the world state.",
-      }),
-      ok: false,
-      status: 409,
-    };
-  }
-
-  if (error.reason === "http_error" && error.safeDeny) {
-    return {
-      error: createErrorResponse({
-        code: "unauthorized",
-        message: "End turn is unavailable for this world.",
-      }),
-      ok: false,
-      status: 403,
-    };
-  }
-
-  if (error.reason === "transition_failed") {
-    return createTransitionFailedResult();
-  }
-
-  if (error.reason === "running_transition") {
-    return createRunningTransitionResult();
-  }
-
-  return createTransitionPersistenceUnavailableResult();
-}
-
-function transitionInputResultFromStateFetchError(
-  error: SupabaseStateFetchError,
-): EndTurnBasicTransitionInputResult {
-  if (error.reason === "missing_world") {
-    return {
-      error: createErrorResponse({
-        code: "end_turn_world_not_found",
-        message: "World is unavailable.",
-      }),
-      ok: false,
-      status: 404,
-    };
-  }
-
-  if (error.reason === "http_error" && error.safeDeny) {
-    return {
-      error: createErrorResponse({
-        code: "unauthorized",
-        message: "End turn is unavailable for this world.",
-      }),
-      ok: false,
-      status: 403,
-    };
-  }
-
-  return createTransitionStateUnavailableResult();
-}
-
-function resultFromSupabaseAuthorizationFetchError(
-  error: SupabaseAuthorizationFetchError,
-): EndTurnBasicAuthorizationResult {
-  if (error.safeDeny) {
-    return createAuthorizationErrorResult();
-  }
-
-  return createAuthContextUnavailableResult();
-}
-
-function parseWorldCalendarConfig(value: unknown): TurnCalendarConfig | null {
-  if (
-    !isRecord(value) ||
-    !hasOnlyExpectedFields(value, expectedCalendarConfigFields)
-  ) {
-    return null;
-  }
-
-  const weekdays = parseCalendarWeekdays(value.weekdays);
-  const months = parseCalendarMonths(value.months);
-
-  if (weekdays === null || months === null) {
-    return null;
-  }
-
-  const startingDayOfMonth = value.startingDayOfMonth;
-  const startingMonthIndex = value.startingMonthIndex;
-  const startingWeekdayOffset = value.startingWeekdayOffset;
-  const startingYear = value.startingYear;
-  const dateFormatTemplate = value.dateFormatTemplate;
-
-  if (
-    !isNonnegativeInteger(startingMonthIndex) ||
-    startingMonthIndex >= months.length ||
-    !isPositiveInteger(startingDayOfMonth) ||
-    !isInteger(startingYear) ||
-    !isNonnegativeInteger(startingWeekdayOffset) ||
-    startingWeekdayOffset >= weekdays.length ||
-    typeof dateFormatTemplate !== "string" ||
-    dateFormatTemplate.trim().length === 0 ||
-    !/\{(?:weekday|month|day|year)\}/.test(dateFormatTemplate) ||
-    /\{(?!weekday\}|month\}|day\}|year\})[^{}]+\}/.test(dateFormatTemplate)
-  ) {
-    return null;
-  }
-
-  const startingMonth = months[startingMonthIndex];
-
-  if (
-    startingMonth === undefined ||
-    startingDayOfMonth > startingMonth.dayCount
-  ) {
-    return null;
-  }
-
-  return {
-    months,
-    startingDayOfMonth,
-    startingMonthIndex,
-    startingWeekdayOffset,
-    startingYear,
-    weekdays,
-    dateFormatTemplate,
-  };
-}
-
-function parseCalendarWeekdays(
-  value: unknown,
-): TurnCalendarConfig["weekdays"] | null {
-  if (!Array.isArray(value) || value.length === 0) {
-    return null;
-  }
-
-  const weekdays: TurnCalendarConfig["weekdays"] = [];
-
-  for (const [weekdayIndex, weekday] of value.entries()) {
-    if (
-      !isRecord(weekday) ||
-      !hasOnlyExpectedFields(weekday, expectedCalendarWeekdayFields) ||
-      weekday.index !== weekdayIndex ||
-      typeof weekday.name !== "string" ||
-      weekday.name.trim().length === 0
-    ) {
-      return null;
-    }
-
-    weekdays.push({
-      index: weekday.index,
-      name: weekday.name,
-    });
-  }
-
-  return weekdays;
-}
-
-function parseCalendarMonths(
-  value: unknown,
-): TurnCalendarConfig["months"] | null {
-  if (!Array.isArray(value) || value.length === 0) {
-    return null;
-  }
-
-  const months: TurnCalendarConfig["months"] = [];
-
-  for (const [monthIndex, month] of value.entries()) {
-    if (
-      !isRecord(month) ||
-      !hasOnlyExpectedFields(month, expectedCalendarMonthFields) ||
-      month.index !== monthIndex ||
-      typeof month.name !== "string" ||
-      month.name.trim().length === 0 ||
-      !isPositiveInteger(month.dayCount)
-    ) {
-      return null;
-    }
-
-    months.push({
-      dayCount: month.dayCount,
-      index: month.index,
-      name: month.name,
-    });
-  }
-
-  return months;
-}
-
-function isSupabaseEndTurnWorldStateRow(
-  value: unknown,
-): value is SupabaseEndTurnWorldStateRow {
-  return (
-    isRecord(value) &&
-    typeof value.id === "string" &&
-    value.id.length > 0 &&
-    typeof value.current_turn_number === "number" &&
-    Number.isInteger(value.current_turn_number) &&
-    value.current_turn_number >= 0 &&
-    (value.status === "active" || value.status === "archived")
-  );
-}
-
-function isNonnegativeInteger(value: unknown): value is number {
-  return typeof value === "number" && Number.isInteger(value) && value >= 0;
-}
-
-function isInteger(value: unknown): value is number {
-  return typeof value === "number" && Number.isInteger(value);
-}
-
-function isPositiveInteger(value: unknown): value is number {
-  return typeof value === "number" && Number.isInteger(value) && value > 0;
-}
-
-function isSupabaseReadinessRow(value: unknown): value is {
-  readonly auto_ready_enabled: boolean;
-  readonly id: string;
-  readonly is_ready_current_turn: boolean;
-} {
-  return (
-    isRecord(value) &&
-    typeof value.id === "string" &&
-    value.id.length > 0 &&
-    typeof value.auto_ready_enabled === "boolean" &&
-    typeof value.is_ready_current_turn === "boolean"
-  );
-}
-
-function isSupabasePersistedTransitionRow(
-  value: unknown,
-): value is SupabasePersistedTransitionRow {
-  return (
-    isRecord(value) &&
-    typeof value.id === "string" &&
-    value.id.length > 0 &&
-    typeof value.world_id === "string" &&
-    value.world_id.length > 0 &&
-    typeof value.from_turn_number === "number" &&
-    Number.isInteger(value.from_turn_number) &&
-    value.from_turn_number >= 0 &&
-    typeof value.to_turn_number === "number" &&
-    Number.isInteger(value.to_turn_number) &&
-    value.to_turn_number === value.from_turn_number + 1 &&
-    typeof value.initiated_by_user_id === "string" &&
-    value.initiated_by_user_id.length > 0 &&
-    typeof value.started_at === "string" &&
-    value.started_at.length > 0 &&
-    (value.status === "completed" ||
-      value.status === "failed" ||
-      value.status === "running")
-  );
-}
-
-function toBasicEndTurnReadinessRow(row: {
-  readonly auto_ready_enabled: boolean;
-  readonly id: string;
-  readonly is_ready_current_turn: boolean;
-}): BasicEndTurnReadinessRow {
-  return {
-    autoReadyEnabled: row.auto_ready_enabled,
-    id: row.id,
-    isReadyCurrentTurn: row.is_ready_current_turn,
-  };
-}
-
-function toEndTurnBasicPersistedTransition(
-  row: SupabasePersistedTransitionRow,
-): EndTurnBasicPersistedTransition {
-  return {
-    fromTurnNumber: row.from_turn_number,
-    id: row.id,
-    initiatedByUserId: row.initiated_by_user_id,
-    startedAt: row.started_at,
-    status: row.status,
-    toTurnNumber: row.to_turn_number,
-    worldId: row.world_id,
-  };
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
 }
 
 const edgeRuntime = getEdgeRuntime();
